# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for managing Akun resources",
        "version": "1.0.0"
    },
    "paths": {
        "/api/akun": {
            "get": {
                "tags": [
                    "Akun"
                ],
                "summary": "Get list of Akun",
                "description": "Display a listing of Akun.",
                "operationId": "beec609b9ae51daef6f12af8d7c611ea",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Akun"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Akun"
                ],
                "summary": "Create a new Akun",
                "description": "Store a new Akun.",
                "operationId": "8216423707756d252a4f28e2f0ebbf89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "kode": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Akun created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Akun"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/akun/{id}": {
            "get": {
                "tags": [
                    "Akun"
                ],
                "summary": "Get an Akun by ID",
                "description": "Display an Akun by ID.",
                "operationId": "507399a2731dcaf28f95368a8f61521f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Akun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Akun not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Akun"
                ],
                "summary": "Update an Akun by ID",
                "description": "Update an existing Akun.",
                "operationId": "55cee838ed1ec965ad407e7fff8b5f94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "kode": {
                                        "type": "integer",
                                        "example": 456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akun updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Akun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Akun not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Akun"
                ],
                "summary": "Delete an Akun by ID",
                "description": "Delete an Akun.",
                "operationId": "5016442ca0c02aa655f562e97a01a6f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Akun deleted"
                    },
                    "404": {
                        "description": "Akun not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Akun": {
                "properties": {
                    "id_akun": {
                        "type": "string",
                        "example": "unique_id_123"
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "kode": {
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}